% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_constancia.R
\name{get_constancia}
\alias{get_constancia}
\title{Obtener constancia de grados y títulos}
\usage{
get_constancia(n_dni = "", nombre = "")
}
\arguments{
\item{n_dni}{character. Un número de DNI otorgado por RENIEC Perú.}

\item{nombre}{character. Apellidos y Nombres de una persona natural.}
}
\value{
character. Información sobre la persona buscada, en formato JSON.
}
\description{
Obtener información sobre los grados y títulos registrados en SUNEDU Perú.
Debe proveerse o un número de DNI o los nombres completos de una persona.
}
\examples{
get_constancia(n_dni = "08192557") # DNI de Superintendente
get_constancia(nombre = "ZEGARRA ROJAS OSWALDO DELFIN") # Nombre de Superintendente
}
