% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_info_from_dni.R
\name{get_info_from_dni}
\alias{get_info_from_dni}
\title{Obtener información desde grados y títulos}
\usage{
get_info_from_dni(n_dni, tipo_info)
}
\arguments{
\item{n_dni}{character. Un número de DNI otorgado por RENIEC Perú.}

\item{tipo_info}{character. Nombre de elemento a obtener.}
}
\value{
character. Información obtenida. Longitud corresponde con registros encontrados.
}
\description{
Permite obtener algún tipo de información específica a partir de un DNI de RENIEC.
}
\examples{
get_info_from_dni(n_dni="08192557", tipo_info = "TIPO_GRADO")
}
