% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mayor_grado_from_dni.R
\name{get_mayor_grado_from_dni}
\alias{get_mayor_grado_from_dni}
\title{Obtener el mayor grado desde un DNI}
\usage{
get_mayor_grado_from_dni(n_dni)
}
\arguments{
\item{n_dni}{character. Un número de DNI otorgado por RENIEC Perú.}
}
\value{
character. El mayor grado académico obtenido por la persona especificada.
}
\description{
Obtener el mayor grado desde un DNI
}
\examples{
get_mayor_grado_from_dni(n_dni = "08192557")
}
