% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_constancia.R
\name{parse_constancia}
\alias{parse_constancia}
\title{Transformar constancia}
\usage{
parse_constancia(constancia)
}
\arguments{
\item{constancia}{character. Constancia de grados y títulos en formato JSON.}
}
\value{
Un data.frame que contiene la información de una o más constancias de grados y títulos
}
\description{
Convierte una constancia de formato JSON a un data.frame válido
}
\examples{
parse_constancia(get_constancia(n_dni="08192557"))
}
